project-structure/utils/ai_integration.py
import os
import json
import numpy as np
from PIL import Image

# Placeholder for your actual AI model integration
# This is where you would import and use your existing model

def preprocess_image(image_path):
    """
    Preprocess the image for the AI model
    Replace this with your actual preprocessing logic
    """
    try:
        img = Image.open(image_path)
        img = img.resize((224, 224))  # Resize to model input size
        img_array = np.array(img) / 255.0  # Normalize
        
        # Add more preprocessing as needed for your model
        
        return img_array
    except Exception as e:
        print(f"Error preprocessing image: {e}")
        return None

def predict_recyclability(image_path):
    """
    Predict if the item in the image is recyclable
    Replace this with your actual model inference logic
    """
    try:
        # Preprocess the image
        processed_image = preprocess_image(image_path)
        
        if processed_image is None:
            return {
                'recyclable': False,
                'confidence': 0.0,
                'error': 'Failed to process image'
            }
        
        # PLACEHOLDER: Replace with your actual model prediction
        # This is where you would load and use your trained model
        # For example:
        # model = load_your_model()
        # prediction = model.predict(processed_image)
        
        # For demonstration, we're returning a random result
        # Replace this with actual model inference
        import random
        is_recyclable = random.choice([True, False])
        confidence = random.uniform(0.7, 0.99)
        
        materials = ["plastic", "paper", "glass", "metal", "organic"] if is_recyclable else ["mixed waste", "hazardous", "non-recyclable plastic"]
        material = random.choice(materials)
        
        return {
            'recyclable': is_recyclable,
            'confidence': confidence,
            'material': material,
            'recommendations': get_recommendations(is_recyclable, material)
        }
    except Exception as e:
        print(f"Error predicting recyclability: {e}")
        return {
            'recyclable': False,
            'confidence': 0.0,
            'error': str(e)
        }

def get_recommendations(is_recyclable, material):
    """
    Generate recommendations based on prediction
    """
    recommendations = {
        "plastic": [
            "Rinse before recycling",
            "Remove labels if possible",
            "Check local recycling guidelines for plastic types"
        ],
        "paper": [
            "Keep dry and clean",
            "Remove any non-paper attachments",
            "Flatten boxes to save space"
        ],
        "glass": [
            "Rinse thoroughly",
            "Remove lids and caps",
            "Sort by color if required by local recycling"
        ],
        "metal": [
            "Rinse containers",
            "Remove non-metal parts",
            "Crush if possible to save space"
        ],
        "organic": [
            "Compost in garden or use organic waste collection",
            "Keep separate from other recyclables",
            "Consider home composting"
        ],
        "mixed waste": [
            "Try to separate recyclable components",
            "Check if special disposal facilities exist"
        ],
        "hazardous": [
            "Take to specialized collection points",
            "Never mix with regular waste",
            "Check local hazardous waste disposal guidelines"
        ],
        "non-recyclable plastic": [
            "Consider reusing if possible",
            "Look for sustainable alternatives in the future",
            "Dispose in general waste"
        ]
    }
    
    return recommendations.get(material, ["Check local waste management guidelines"])