project-structure/requirements.txt
Flask==2.3.3
pymongo==4.5.0
flask-bcrypt==1.0.1
python-dotenv==1.0.0
flask-login==0.6.2
Pillow==10.0.0
python-magic==0.4.27
Werkzeug==2.3.7
Flask-WTF==1.1.1
email-validator==2.0.0